requirements.txt
selenium
python-dotenv
chromedriver-autoinstaller
python-dateutil 